Make all run outputs visible in AML and improve documentation (#147)
Change the value of RUN_OUTPUTS_DIR_NAME from "run_outputs" to DEFAULT_AML_UPLOAD_DIR which is "outputs". This means that output files are automatically uploaded to blob storage and are also all visible in AzureML.
Add a check at the end of the run method in runner.py that the run_outputs directory is indeed the default upload directory (this will always succeed at the moment) and if not, upload the directory explicitly.
Improve the documentation on the model output directory so that it covers everything that's there.