Rationalize directory structure in plotting and ensemble building (#169)
Reorganize outputs from train and test runs. For details of the new structure, see the changes to building_models.md, which should be read first.

Also:

Pass is_ensemble as an argument to lots of places from the ultimate calling context, so we can be sure when we're running an ensemble model

Add a context manager logging_section, to support named sections in the log to aid in understanding them.
Remove baseline-comparison results after downloading and doing the comparison.

For each training and validation epoch, log the total time taken and the portion of that taken for data loading.

Use shortened split names in box plots so X axis labels can be read (there will be more to do in future PRs to improve plot legibility).

Add a switch --create_plots, default True, to plot_cross_validation.py so that by setting it to False, the script can be run (to generate outliers and statistical test results) in environments where graphics is not available.