Various improvements and clarifications (#175)
Rename ModelType to ModelProcessing, ModelType.SINGLE to ModelProcessing.DEFAULT, and ModelType.ENSEMBLE to ModelType.ENSEMBLE_CREATION, to express more clearly what the roles of these values are; and document the Enum.

Introduce a dataclass ModelAndInfo for use in model loading and adjustment, allowing the relevant code to be refactored, and ensure adjustment (for mixed precision etc) is only attempted once, so model re-training can in principle work.

Extend the semantics of the --restrict_subjects switch so that if one of the fields in the value is "+", discarded members of either or both of the other sets are added to the "+" set. For example "+,,10" means "move all but 10 members of the test set to the training set". Previously this only worked when the number was 0.

Ensemble model aggregation now happens lazily - i.e. individual predictions are now produced by a generator, and the aggregation code ensures only a minimal number of posterior-probability arrays (two) exist at any one time. This is because we can otherwise get (main) memory overflows from a head-and-neck ensemble model because of the large number of structures.

Ensemble models are (since recently) registered by child 0 during training, and single models have always been registered during training. So there should now be no need for runs with is_train=False to register a model. This is implemented here, with a "back door" that if the user supplies the (existing) switch --register_model_only_for_epoch, they must have meant registration to happen, perhaps because it failed for some reason in the training run.

Remove the --model_name and --model_version switches from submit_for_inference.py; they were causing confusion, so now --model_id is the only way to specify the model.

Refactor part of mypy_runner.py (no change to functionality).

Set the dtype of the posterior array in inference to np.float32 (instead of 64), thereby saving lots of memory so that HeadAndNeck models work in the ASH configuration.