Adding mutually exclusive label check to load_labels_from_dataset_source (#454)
Issue

Closes #339

Overview

We need to check that labels are mutually exclusive early, when the labels are loaded from the data source.

Details

load_labels_from_dataset_source builds up the ground truth tensor from individual binary masks. It sets the background mask correctly, but does not check if each pixel is really only in 1 class. If that assumption is violated, model training may later fail when computing the FocalLoss, and when it fails later the error message does not explain the reason well. So we need to check that the labels are mutually exclusive early, when the labels are loaded from the data source.

Some loss functions (e.g. SoftDice) may cope with overlapping labels so we provide a flag for users to turn off the check, but by default if the labels are not mutually exclusive we now throw a ValueError with a suitably explanatory message.