ENH: Improve regression tests (#827)
Closes #740. Updates the lung model regression test to use the latest
parameters and train for a substantial number of steps to ensure
training is progressing as expecting. The small number of epochs and
smaller data subset is used as running a full training run isn't
feasible. The new test runs in < 30 minutes but on real data.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>

Bypassing branch protections for failing PyTest as the test that is failing is `Tests/SSL/test_ssl_containers.py::test_innereye_ssl_container_cifar10_resnet_simclr` as the URL for CIFAR seems to be broken currently (returning 500 server error).