ENH: Adapt HelloWorld to run using the Inference Service (#837)
Closes #696.

Changes the HelloWorld model to only use a single channel, as this is
the format that the [Inference
Service](https://github.com/microsoft/InnerEye-Inference/) expects for
its inputs.

As many tests require the 2-channel data, this PR also creates a new
class, `HelloWorld2Channel`, which inherits from `HelloWorld` but uses 2
channels and can be used for testing. This avoids the pain of having to
alter all unit and regression tests that relied on the 2-channel data
and model.

Both models can be trained on the same data, but now the `HelloWorld`
model can be run by the Inference Service.